Producer wait for response from consumer
This commit includes the following:
Producer waits for 10 seconds to listen back from consumer after the operation is completed from consumer end. This is added as in the next phase, we will use the image(req body param) for inference and we should return the response back to API req when consumer processes the image.